added 'text arithmetic parsing', Overview notebook, and fixed bug
By 'text arithmetic parsing,' I mean I added the ability to specify that the
value of a variable as 'h/3' so that 'h/3' varies with the values of the
variable 'h' but that is always 1/3 the value of 'h'. This allows for custom
embeddings/ chainings to work properly.

By Overview notebook, I mean a notebook that encapsulates all the features
that this software has in store for presentation purposes.

Fixed bug: It turns out, one of the conditionals in the numeric_anneal_schedule
was actually wrong--resulting in the wrong A(t) and B(t) function. Fixed now.