added densify_unitcell method in dictrep class
The purpose of "densify_unitcell" is to heuristically create a maximal length
qubit chain within a unit cell. The method takes as argument an initial qubit
(which asks the user to give the lowest # qubit in a unit cell), an hweight,
and a Jweight. It then creates the chain (that's never broken) of maximal
length with each qubit getting an hweight and coupling a Jweight.