moved densify_unitcell to dwavetools + working on 3Qubit problem
It makes more sense for the utility that densify_unitcell is to be in the
dwavetools python file rather than a class method.

In regards to the 3Qubit problem with huge chaining, there is some code in
3Qubit_BigChain.ipynb that begins to work on this. In particular, it tries to
create dense rectangles connected as densely as possible. There appears to be
a bug with adding the second qubit, however.