dense chain function appears to be working
the dense_connect_2000Q function in MSU_Research/testing.ipynb appears to be
working. The details of how it works are self-explanatory in the code comments.
Further testing should include integrating networkx outputs with gephi and
visualizing created chains in gephi w.r.t. working qubits/ couplers on the
Chimera.