major updates to ErrorCorrection routine and some Malay testing
Updates to FREM can be summarzied in MSU_testing/EC_Tests2.ipyn. The idea is
that a numerical forward, reverse, and frem anneal have been added as calls
to the DictRep class (associated changes to dictrep.py and dwavetools.py).

Malay testing was really just ensuring that my ground-states/ numerical
forward annealing are working properly and to give example testing code. This
will be useful later when it comes to FREM testing, however (somewhat).

Need-to-dos:
1) Create function that finds most probable state in computational basis
after an anneal to put into FREM algorithm.
2) Finish FREM algorithm.
3) Generate a very frustrated Hamiltonian to perform tests on.
4) Find the largest Hamiltonian that QuTip can handle on my own machine.
5) Add compatibility with storing data/ parameter sweeping/ and easy
state plotting with numerical anneals.
6) Do some FREM testing, especially on the frustrated system.